Displays memberships in 'general info' editor component
