+++ services/admin/src/routes/object/[prefix]/[noid].svelte
@@ -14,7 +14,11 @@
         ].join("/");
         const response = await context.lapin.query("accessObject.get", id);
         const serverObject = AccessObject.parse(response);
-        return { props: { serverObject } };
+        const membership = await context.lapin.query(
+          "accessObject.getMembership",
+          id
+        );
+        return { props: { serverObject, membership } };
       }
       return { props: { error: "Could not find prefix or noid in url." } };
     } catch (e) {
@@ -30,6 +34,7 @@
    * The object is given to the page from the module above.
    */
   import { AccessObject } from "@crkn-rcdr/access-data";
+  import type { Membership } from "@crkn-rcdr/access-data";
   import Editor from "$lib/components/access-objects/Editor.svelte";
   import NotificationBar from "$lib/components/shared/NotificationBar.svelte";
   import Loading from "$lib/components/shared/Loading.svelte";
@@ -40,13 +45,18 @@
   export let serverObject: AccessObject;
 
   /**
+   * Membership record for this object.
+   */
+  export let membership: Membership;
+
+  /**
    * @type {string} An error message insdicating what went wrong.
    */
   export let error: string;
 </script>
 
 {#if serverObject}
-  <Editor bind:serverObject />
+  <Editor bind:serverObject {membership} />
 {:else if error}
   <br />
   <div class="wrapper">
